pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "beautifulsoup4>=4.14.2",
    "cartopy>=0.25.0",
    "fastapi[standard]>=0.121.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "rasterio>=1.4.3",
    "requests>=2.32.5",
    "rioxarray>=0.19.0",
    "scipy>=1.16.3",
    "sqlalchemy>=2.0.44",
    "torch",
]
